+++ R/package.R
@@ -15,5 +15,7 @@
 .onLoad <- function(...) {
   S7::method(print, s7_interface) <- .print_s7_interface
   S7::method(print, s7_trait) <- .print_s7_trait
+  S7::method(with, s7_interface) <- .with_s7_interface
+  S7::method(with, s7_trait) <- .with_s7_trait
   S7::methods_register()
 }